# Microservices.IoT.Fridge/swagger JSONs/management server swagger definition.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Microservices.IoT.ManagementConsole.RestAPI",
    "version": "1.0"
  },
  "paths": {
...
    "/api/FoodItemsSimulation/{id}/open": {
      "patch": {
        "tags": [
          "FoodItemsSimulation"
        ],
        "summary": "Opens the food item's packaging.\r\n<br>Does nothing if the food packaging is already open.</br>",
        "operationId": "OpenFoodItemPackage",
        "parameters": [
          {
            "name": "ID",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/FoodItemsSimulation/{id}": {
      "post": {
        "tags": [
          "FoodItemsSimulation"
        ],
        "summary": "Simulates consuming of a part of a food item.\r\n<br>Decreases the weight of a particular food item by percentage amountPercent.</br><br>A food item has initial weight and current weight in grams.</br><br>If the weight reaches 0, the food item disappears (is deleted).</br><br>Opens the food item's package if this is first consumption</br>",
        "operationId": "ConsumeFoodItem",
        "parameters": [
          {
            "name": "ID",
            "in": "path",
            "description": "food ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "amountPercent",
            "in": "query",
            "description": "must be between 1 and 100",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "409": {
            "description": "Conflict",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "400": {
            "description": "If amountPercent is not a number between 1 and 100"
          }
        }
      }
    },
...
    "/api/FridgeSensorsManagement/{name}": {
      "get": {
        "tags": [
          "FridgeSensorsManagement"
        ],
        "summary": "= microservice process for this fridge is running",
        "operationId": "IsSensorMicroserviceRunning",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "boolean"
                }
              },
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              },
              "text/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        }
      }
    },
    "/api/FridgeSensorsManagement/{name}/port": {
      "get": {
        "tags": [
          "FridgeSensorsManagement"
        ],
        "summary": "Returns the port on which a sensor microservice process is running if it is tracked\r\n<br>`null` if the service is not running (or is no tracked)</br>",
        "operationId": "GetRunningSensorMicroservicePort",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              },
              "application/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              },
              "text/json": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        }
      }
    },
    "/api/FridgeSensorsManagement/{name}/start": {
      "post": {
        "tags": [
          "FridgeSensorsManagement"
        ],
        "summary": "Starts a new instance of a microservice with a given name to listen on port port and adds it to name service-",
        "operationId": "StartSensorMicroservice",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "port",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "showWindow",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "If a service for name was sucessfully started"
          },
          "409": {
            "description": "If a service for name is already running or if port is already occupied",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "If a service for name fails to start",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/FridgeSensorsManagement/{name}/stop": {
      "post": {
        "tags": [
          "FridgeSensorsManagement"
        ],
        "summary": "Stops a sensor microservice running for defined name",
        "operationId": "StopSensorMicroservice",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "boolean"
                }
              },
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              },
              "text/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        }
      }
    },
    "/api/FridgeSensorsManagement/sensorProcessExecutablePath": {
      "put": {
        "tags": [
          "FridgeSensorsManagement"
        ],
        "summary": "Sets path to a file which will be executed as a fridge sensor microservice.",
        "operationId": "SetSensorProcessExecutablePath",
        "parameters": [
          {
            "name": "path",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "get": {
        "tags": [
          "FridgeSensorsManagement"
        ],
        "summary": "Gets the path to the executable which is used as a fridge sensor microservice process.",
        "operationId": "GetSensorProcessExecutablePath",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      }
    }
  }
}